Adds smarter reordering rules.
We can reorder two properties if one them is shorthand and the other
is granular but they have the same value.

Note: it may be improved even further with #429 in place. At that time
we may compare exact values, e.g. `margin:0 1px` is currently not
swappable with `margin-top:0` but it should.