Fixes #397 - support for source map's sourcesContent property.
When handling input source map it checks whether sourcesContent is present
and if it is so then it's reused in the output source map.

Adds `sourceMapInlineSources` / `--source-map-inline-source` switches to
control whether an inlined source map is created or not.

In case an input source map with a `sourcesContent` field is provided
then all sources from that source map are carried over to the output source map.