Fixes #498 - restructuring content with flexbox properties.
Flexbox is quite tricky as there are many different properties which
can't be reordered. Instead of implementing fine grained rules for all
of them we simply block flexbox from being restructured.

Potentially we should revisit it at some point to find out if restructuring
groups of properties would be a good idea.