Fixes #543 - better "comment in body" handling.
So instead of keeping comments in a property list we get them out
of every property.

It is a bit tricky as comments escaped at the beginning need to be
tracked and saved, but ones from the middle or end of the value has
to be saved for later (see `innerProperties`) and tracked after the
property even if they are moved before it.

See tests for a better picture.