Adds semantic merging (off by default).
The main obstacle on reordering and merging declarations is a fact that
arbitrary classes can be applied to an element and at CSS level we don't
know if that's the case or not.

However with semantic merging mode on we trust CSS author knows what she is doing.

This commit is just a start (see #588) of a journey. It is and will always
be turned off by default as it requires certain effort from stylesheets'
author.

So far plain class selectors and some BEM basics are supported.