Disables URL rebasing by default.
If you give clean-css a `rebaseTo` option this still counts as turning rebasing
ON (preserved for some backwards compatibility).

Please note that when processing imports, relative URLs in imported files are
rebased according to file which required the imported file. This is probably
what you want as otherwise URLs will break.